# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mediate Healthcare MR App",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/territories/my-territories": {
      "get": {
        "tags": [
          "Territories"
        ],
        "summary": "Get logged-in user's assigned territories",
        "description": "Fetch active territories assigned to current MR / Manager.",
        "operationId": "get_my_territories_api_v1_territories_my_territories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TerritoryResponse"
                  },
                  "type": "array",
                  "title": "Response Get My Territories Api V1 Territories My Territories Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/territories": {
      "get": {
        "tags": [
          "Territories"
        ],
        "summary": "List all territories",
        "description": "List territories with search and state filter.",
        "operationId": "list_territories_api_v1_territories_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "state_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State Id"
            }
          },
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Active Only"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TerritoryResponse"
                  },
                  "title": "Response List Territories Api V1 Territories Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Territories"
        ],
        "summary": "Create territory (Admin)",
        "description": "Create a new territory and assign areas.",
        "operationId": "create_territory_api_v1_territories_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerritoryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerritoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/territories/{territory_id}": {
      "get": {
        "tags": [
          "Territories"
        ],
        "summary": "Get territory details by ID",
        "description": "Get territory by ID.",
        "operationId": "get_territory_api_v1_territories__territory_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "territory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Territory Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerritoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Territories"
        ],
        "summary": "Update territory (Admin)",
        "description": "Update territory metadata and areas.",
        "operationId": "update_territory_api_v1_territories__territory_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "territory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Territory Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerritoryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerritoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/territories/{territory_id}/assign": {
      "post": {
        "tags": [
          "Territories"
        ],
        "summary": "Assign user to territory",
        "description": "Assign an MR or Manager to a territory.",
        "operationId": "assign_user_to_territory_api_v1_territories__territory_id__assign_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "territory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Territory Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTerritoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerritoryAssignmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/territories/{territory_id}/assign/{user_id}": {
      "delete": {
        "tags": [
          "Territories"
        ],
        "summary": "Unassign user from territory",
        "description": "Remove user from territory.",
        "operationId": "unassign_user_from_territory_api_v1_territories__territory_id__assign__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "territory_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Territory Id"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Unassign User From Territory Api V1 Territories  Territory Id  Assign  User Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AreaResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "city_id": {
            "type": "integer",
            "title": "City Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "pincode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pincode"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "city_id",
          "name",
          "is_active"
        ],
        "title": "AreaResponse",
        "description": "Area / Locality response."
      },
...
      "AssignTerritoryRequest": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "AssignTerritoryRequest",
        "description": "Request to assign an MR or Manager to a Territory."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TerritoryAssignmentResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "user_name": {
            "type": "string",
            "title": "User Name"
          },
          "territory_id": {
            "type": "integer",
            "title": "Territory Id"
          },
          "territory_name": {
            "type": "string",
            "title": "Territory Name"
          },
          "assigned_at": {
            "type": "string",
            "format": "date-time",
            "title": "Assigned At"
          },
          "unassigned_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unassigned At"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "user_name",
          "territory_id",
          "territory_name",
          "assigned_at",
          "is_active"
        ],
        "title": "TerritoryAssignmentResponse",
        "description": "Record of user territory allocation."
      },
      "TerritoryCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name"
          },
          "code": {
            "type": "string",
            "maxLength": 50,
            "title": "Code"
          },
          "headquarters": {
            "type": "string",
            "maxLength": 100,
            "title": "Headquarters"
          },
          "state_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "area_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Area Ids"
          }
        },
        "type": "object",
        "required": [
          "name",
          "code",
          "headquarters"
        ],
        "title": "TerritoryCreate",
        "description": "Payload to create a Territory."
      },
      "TerritoryResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "headquarters": {
            "type": "string",
            "title": "Headquarters"
          },
          "state_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "areas": {
            "items": {
              "$ref": "#/components/schemas/AreaResponse"
            },
            "type": "array",
            "title": "Areas",
            "default": []
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "code",
          "headquarters",
          "is_active",
          "created_at"
        ],
        "title": "TerritoryResponse",
        "description": "Territory response with mapped areas."
      },
      "TerritoryUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "headquarters": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Headquarters"
          },
          "state_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "area_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Ids"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "TerritoryUpdate",
        "description": "Payload to update a Territory."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}